+++ src/app/components/icon/simple-status-icon.jsx
@@ -94,4 +94,8 @@ SimpleStatusIcon.propTypes = {
 	style: React.PropTypes.object
 };
 
+SimpleStatusIcon.defaultProps = {
+	tooltipPosition: 'top-center'
+};
+
 module.exports = SimpleStatusIcon;